Refactor asset spritesheet modules
