+++ src/heart/assets/loader.py
@@ -5,16 +5,17 @@ from typing import Any
 
 import pygame
 
+from heart.assets.animation import Animation
 from heart.assets.cache import AssetCache
-from heart.utilities.env import (AssetCacheStrategy, Configuration,
-                                 SpritesheetFrameCacheStrategy)
+from heart.assets.spritesheet import Spritesheet
+from heart.utilities.env import AssetCacheStrategy, Configuration
 
 DEFAULT_FONT_SIZE = 10
 
 
 class Loader:
     _image_cache: AssetCache[Path, pygame.Surface] | None = None
-    _spritesheet_cache: AssetCache[Path, "spritesheet"] | None = None
+    _spritesheet_cache: AssetCache[Path, Spritesheet] | None = None
     _metadata_cache: AssetCache[Path, dict[str, Any]] | None = None
 
     @classmethod
@@ -26,7 +27,7 @@ class Loader:
         return Configuration.asset_cache_max_entries()
 
     @classmethod
-    def _get_spritesheet_cache(cls) -> AssetCache[Path, "spritesheet"] | None:
+    def _get_spritesheet_cache(cls) -> AssetCache[Path, Spritesheet] | None:
         strategy = cls._resolve_cache_strategy()
         if strategy not in {AssetCacheStrategy.SPRITESHEETS, AssetCacheStrategy.ALL}:
             return None
@@ -86,21 +87,21 @@ class Loader:
         return loaded
 
     @classmethod
-    def load_spirtesheet(cls, path: str | PathLike[str]) -> "spritesheet":
+    def load_spirtesheet(cls, path: str | PathLike[str]) -> Spritesheet:
         resolved_path = cls._resolve_path(path)
         cache = cls._get_spritesheet_cache()
         if cache is not None:
             cached = cache.get(resolved_path)
             if cached is not None:
                 return cached
-        loaded = spritesheet(resolved_path)
+        loaded = Spritesheet(resolved_path)
         if cache is not None:
             cache.set(resolved_path, loaded)
         return loaded
 
     @classmethod
-    def load_animation(cls, path: str | PathLike[str]) -> "Animation":
-        return Animation(cls._resolve_path(path), 100)
+    def load_animation(cls, path: str | PathLike[str]) -> Animation:
+        return Animation(cls.load_spirtesheet(path), 100)
 
     @classmethod
     def load_font(
@@ -120,104 +121,3 @@ class Loader:
         if cache is not None:
             cache.set(resolved_path, payload)
         return dict(payload)
-
-
-# https://www.pygame.org/wiki/Spritesheet
-# I copied this from here it is kinda meh lol
-class spritesheet(object):
-    def __init__(self, filename: str):
-        path = Path(filename)
-        if not path.exists():
-            raise ValueError(f"'{path}' does not exist.")
-
-        if not path.is_file():
-            raise ValueError(f"'{path}' is not a file.")
-
-        with path.open("rb") as f:
-            self.sheet = pygame.image.load(f).convert_alpha()
-        self._frame_cache: dict[tuple[int, int, int, int], pygame.Surface] = {}
-        self._scaled_cache: dict[tuple[int, int, int, int, int, int], pygame.Surface] = {}
-
-    def get_size(self):
-        return self.sheet.get_size()
-
-    def image_at(self, rectangle):
-        rect = pygame.Rect(rectangle)
-        cache_key = (rect.x, rect.y, rect.width, rect.height)
-        strategy = Configuration.spritesheet_frame_cache_strategy()
-        if strategy in {
-            SpritesheetFrameCacheStrategy.FRAMES,
-            SpritesheetFrameCacheStrategy.SCALED,
-        }:
-            cached = self._frame_cache.get(cache_key)
-            if cached is not None:
-                return cached
-
-        image = pygame.Surface(rect.size, pygame.SRCALPHA)
-        image.blit(self.sheet, (0, 0), rect)
-        if strategy in {
-            SpritesheetFrameCacheStrategy.FRAMES,
-            SpritesheetFrameCacheStrategy.SCALED,
-        }:
-            self._frame_cache[cache_key] = image
-        return image
-
-    def image_at_scaled(self, rectangle, size: tuple[int, int]) -> pygame.Surface:
-        rect = pygame.Rect(rectangle)
-        width, height = size
-        cache_key = (rect.x, rect.y, rect.width, rect.height, width, height)
-        strategy = Configuration.spritesheet_frame_cache_strategy()
-        if strategy == SpritesheetFrameCacheStrategy.SCALED:
-            cached = self._scaled_cache.get(cache_key)
-            if cached is not None:
-                return cached
-
-        image = self.image_at(rect)
-        scaled = pygame.transform.scale(image, (width, height))
-        if strategy == SpritesheetFrameCacheStrategy.SCALED:
-            self._scaled_cache[cache_key] = scaled
-        return scaled
-
-    def images_at(self, rects):
-        "Loads multiple images, supply a list of coordinates"
-        return [self.image_at(rect) for rect in rects]
-
-    def load_strip(self, rect, image_count):
-        "Loads a strip of images and returns them as a list"
-        tups = [
-            (rect[0] + rect[2] * x, rect[1], rect[2], rect[3])
-            for x in range(image_count)
-        ]
-        return self.images_at(tups)
-
-
-class Animation:
-    def __init__(self, file_path: str, width: int) -> None:
-        self.spritesheet = Loader.load_spirtesheet(file_path)
-
-        image_size = self.spritesheet.sheet.get_size()
-        self.number_of_frames = image_size[0] / width
-
-        self.key_frames = [
-            (width * i, 0, width, image_size[1])
-            for i in range(int(self.number_of_frames))
-        ]
-        self.current_frame = 0
-        self.ms_since_last_update = None
-        self.ms_per_frame = 25
-
-    def step(self, window, clock):
-        if (
-            self.ms_since_last_update is None
-            or self.ms_since_last_update > self.ms_per_frame
-        ):
-            if self.current_frame >= self.number_of_frames - 1:
-                self.current_frame = 0
-            else:
-                self.current_frame += 1
-
-            self.ms_since_last_update = 0
-
-        image = self.spritesheet.image_at(self.key_frames[self.current_frame])
-        self.ms_since_last_update += clock.get_time()
-        return image
